requirements.txt
Django>=3.2.8,<=3.3.0
djangorestframework>=3.12.4,<=3.13.0
psycopg2>=2.9.1,<=3.0.0
Pillow>=8.4.0,<=8.5.0

flake8>=4.0.0,<=4.1.0
black==19.3b0